Implement the u command
Undoes the effect of the last command that modified anything in the
buffer and restores the current address to what it was before the
command. The global commands 'g', 'G', 'v', and 'V' are treated as a
single command by undo. 'u' is its own inverse; it can undo only the
last command.

Note the introduction of ∆∆ and __, which hold the undo buffer and
undo current line, respectively.